refactor: address review comments on LinearGaussianSCM (#3468)
- Rename LinearGaussian -> LinearGaussianSCM (Ankur, Runak)
- Rename linear_gaussian.py -> linear_gaussian_scm.py
- Convert all docstrings to numpydoc format
- Remove :meth: Sphinx directives not used in codebase
- Expand _build_model docstring with full Parameters section
- Move tests to separate file test_linear_gaussian_scm.py (Ankur)
- Remove linear_gaussian from ALL_DATASETS in test_datasets.py
